+++ example/lib/widgets/examples/examples.dart
@@ -13,3 +13,4 @@ export 'custom_styling_example_widget.dart';
 export 'read_only_example_widget.dart';
 export 'minimal_example_widget.dart';
 export 'tickmark_positioning_example_widget.dart';
+export 'segment_display_example_widget.dart';