+++ lib/src/multi_thumb_slider.dart
@@ -1,5 +1,6 @@
 export 'constants.dart';
 export 'value_type_handler.dart';
 export 'position_calculator.dart';
+export 'segment_calculator.dart';
 export 'widgets/widgets.dart';
 export 'custom_multi_thumb_slider.dart';