+++ lib/src/widgets/widgets.dart
@@ -3,3 +3,4 @@ export 'tickmark_label_widget.dart';
 export 'range_segment_widget.dart';
 export 'tooltip_widget.dart';
 export 'thumb_widget.dart';
+export 'segment_display_widget.dart';